feat: code for the buttons and minor changes
