+++ StoreManager/StoreManager.csproj
@@ -202,29 +202,17 @@
     <Compile Include="UsrCtrlCashiering.Designer.cs">
       <DependentUpon>UsrCtrlCashiering.cs</DependentUpon>
     </Compile>
-    <Compile Include="UsrCtrlInventory.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="UsrCtrlInventory.Designer.cs">
-      <DependentUpon>UsrCtrlInventory.cs</DependentUpon>
-    </Compile>
     <Compile Include="UsrCtrlAnalytics.cs">
       <SubType>UserControl</SubType>
     </Compile>
     <Compile Include="UsrCtrlAnalytics.Designer.cs">
       <DependentUpon>UsrCtrlAnalytics.cs</DependentUpon>
     </Compile>
-    <Compile Include="UsrCtrlInventory2Resize.cs">
+    <Compile Include="UsrCtrlInventory.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="UsrCtrlInventory2Resize.Designer.cs">
-      <DependentUpon>UsrCtrlInventory2Resize.cs</DependentUpon>
-    </Compile>
-    <Compile Include="UsrCtrlInventory2.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="UsrCtrlInventory2.Designer.cs">
-      <DependentUpon>UsrCtrlInventory2.cs</DependentUpon>
+    <Compile Include="UsrCtrlInventory.Designer.cs">
+      <DependentUpon>UsrCtrlInventory.cs</DependentUpon>
     </Compile>
     <Compile Include="UsrCtrlStaff.cs">
       <SubType>UserControl</SubType>
@@ -258,17 +246,11 @@
     <EmbeddedResource Include="UsrCtrlCashiering.resx">
       <DependentUpon>UsrCtrlCashiering.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="UsrCtrlInventory.resx">
-      <DependentUpon>UsrCtrlInventory.cs</DependentUpon>
-    </EmbeddedResource>
     <EmbeddedResource Include="UsrCtrlAnalytics.resx">
       <DependentUpon>UsrCtrlAnalytics.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="UsrCtrlInventory2Resize.resx">
-      <DependentUpon>UsrCtrlInventory2Resize.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="UsrCtrlInventory2.resx">
-      <DependentUpon>UsrCtrlInventory2.cs</DependentUpon>
+    <EmbeddedResource Include="UsrCtrlInventory.resx">
+      <DependentUpon>UsrCtrlInventory.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="UsrCtrlStaff.resx">
       <DependentUpon>UsrCtrlStaff.cs</DependentUpon>
@@ -309,6 +291,7 @@
     <Content Include="Resources\search_icon.png" />
     <None Include="Resources\tag.png" />
     <None Include="Resources\shoes.png" />
+    <Content Include="Resources\staff.png" />
     <Content Include="Resources\upload_icon.png" />
   </ItemGroup>
   <ItemGroup>
